+++ packages/adapter-vercel/files/entry.js
@@ -1,4 +1,4 @@
-import { getRawBody } from '@sveltejs/kit/http'; // eslint-disable-line import/no-unresolved
+import { getRawBody } from '@sveltejs/kit/node'; // eslint-disable-line import/no-unresolved
 import '@sveltejs/kit/install-fetch'; // eslint-disable-line import/no-unresolved
 
 // TODO hardcoding the relative location makes this brittle
