+++ packages/kit/rollup.config.js
@@ -42,7 +42,7 @@ export default [
 		input: {
 			cli: 'src/cli.js',
 			ssr: 'src/runtime/server/index.js',
-			http: 'src/core/http/index.js',
+			node: 'src/core/node/index.js',
 			'install-fetch': 'src/install-fetch.js'
 		},
 		output: {
