+++ packages/kit/src/core/dev/index.js
@@ -10,7 +10,7 @@ import create_manifest_data from '../../core/create_manifest_data/index.js';
 import { create_app } from '../../core/create_app/index.js';
 import { rimraf } from '../filesystem/index.js';
 import { respond } from '../../runtime/server/index.js';
-import { getRawBody } from '../http/index.js';
+import { getRawBody } from '../node/index.js';
 import { copy_assets, get_no_external, resolve_entry } from '../utils.js';
 import svelte from '@sveltejs/vite-plugin-svelte';
 import { get_server } from '../server/index.js';
